backend/requirements.txt
# Core FastAPI and async support
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0

# Database
asyncpg==0.29.0
psycopg2-binary==2.9.9
sqlalchemy==2.0.25
alembic==1.13.1

# Redis
redis[hiredis]==5.0.1
aioredis==2.0.1

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Google Gemini (for Phase 4)
google-generativeai==0.3.2

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
httpx==0.26.0
faker==22.6.0

# Utilities
python-dotenv==1.0.0
python-dateutil==2.8.2
pytz==2023.3

# Monitoring
prometheus-client==0.19.0

# Data validation
email-validator==2.1.0
phonenumbers==8.13.27

# Development tools
black==23.12.1
flake8==7.0.0
mypy==1.8.0
